fix: refactor verifyHandler to streamline imports and improve readability
